Refactor everything into a single component
